修复删除JavaScript进程，修改依赖重新安装逻辑

+++ back/services/cron.ts
@@ -245,16 +245,20 @@ export default class CronService {
       } else {
         return;
       }
-      const pids = pid.match(/\(\d+/g);
+      let pids = pid.match(/\(\d+/g);
       const killLogs = [];
-      for (const id of pids) {
-        const c = `kill -9 ${id.slice(1)}`;
-        const { stdout, stderr } = await execAsync(c);
-        if (stderr) {
-          killLogs.push(stderr);
-        }
-        if (stdout) {
-          killLogs.push(stdout);
+      if (pids && pids.length > 0) {
+        // node 执行脚本时还会有10个子进程，但是ps -ef中不存在，所以截取前三个
+        pids = pids.slice(0, 3);
+        for (const id of pids) {
+          const c = `kill -9 ${id.slice(1)}`;
+          const { stdout, stderr } = await execAsync(c);
+          if (stderr) {
+            killLogs.push(stderr);
+          }
+          if (stdout) {
+            killLogs.push(stdout);
+          }
         }
       }
       return killLogs.length > 0 ? JSON.stringify(killLogs) : '';
